fix: cast to float32 before sigmoid in all loss functions for AMP safety
All loss functions that call torch.sigmoid() on raw logits now cast to
float32 first (input.float()) to prevent NaN gradients when AMP runs the
forward pass in float16. In float16, sigmoid can underflow/overflow at
extreme logit values, producing 0.0 or 1.0 exactly, which then causes
division-by-zero or log(0) in downstream Tversky/Dice computations.

Affected functions:
- PartialTverskyLoss._per_channel_tversky_loss()
- BalancedSoftmaxTverskyLoss.forward()
- DiceBCELoss.forward() (in loss_zoo.py)
- BoundaryWeightedTverskyLoss.forward()

FocalTverskyLoss and AsymmetricUnifiedFocalLoss already had this fix.

+++ training/losses/boundary_loss.py
@@ -130,7 +130,7 @@ class BoundaryWeightedTverskyLoss(nn.Module):
         mask = self._annotation_mask
         self._annotation_mask = None
 
-        pred = torch.sigmoid(input)
+        pred = torch.sigmoid(input.float())  # float32 for AMP safety
         target = target.float()
         spatial_dims = tuple(range(2, input.ndim))
